split numeric into another submodule
